Updated Hamiltonians module, moved hubbard into it
Moved Hubbard definition into latticemodels.jl from pools.jl, added the documentation to docs
