+++ src/base/pools.jl
@@ -3,12 +3,10 @@
 #= TO DO: tile operators by removing leading and trailing I's first. =#
 
 module Pools
-    import LinearAlgebra: I
 
     import PauliOperators
     import PauliOperators: FixedPhasePauli, Pauli, ScaledPauli
     import PauliOperators: ScaledPauliVector, PauliSum
-    import PauliOperators: clip!, jordan_wigner
 
     import ..MyPauliOperators: otimes, ≈
     import PauliOperators: ⊗, ≈
@@ -52,87 +50,6 @@ module Pools
             -ScaledPauli(Pauli(n; X=[j], Y=[i,k,l])),
             -ScaledPauli(Pauli(n; X=[i], Y=[j,k,l])),
         ]
-    end
-
-
-
-
-    """
-        hubbard_jw(graph::Array{T,2}, U, t)
-
-    A Hubbard Hamiltonian in the Jordan-Wigner basis.
-
-    Copied shamelessly from Diksha's ACSE repository.
-
-    # Parameters
-    - `graph`: an adjacency matrix identifying couplings. Must be symmetric.
-    - `U`: Coulomb interaction for all sites
-    - `t`: hopping energy for all couplings
-
-    # Returns
-    - `PauliOperators.PauliSum`: the Hamiltonian
-
-    """
-    function hubbard_hamiltonian(graph::Matrix{T}, U, t) where T
-        Ni, Nj = size(graph)
-        Ni == Nj || throw(DimensionMismatch)
-        Norb = Ni
-        N = 2*Norb
-        H = PauliSum(N)
-        for i in 1:Norb
-            ia = 2*i-1
-            ib = 2*i
-            for j in i+1:Norb
-                ja = 2*j-1
-                jb = 2*j
-                abs(graph[i,j]) > 1e-16 || continue
-
-                tij = jordan_wigner(ia, N)*jordan_wigner(ja,N)'
-                tij += jordan_wigner(ib, N)*jordan_wigner(jb,N)'
-                tij += adjoint(tij)
-                clip!(tij)
-                sum!(H,t*tij)
-            end
-
-            ni = jordan_wigner(ia, N)*jordan_wigner(ia,N)'*jordan_wigner(ib, N)*jordan_wigner(ib,N)'
-            sum!(H,U*ni)
-        end
-        return H
-    end
-
-    """
-        hubbard_hamiltonian(L::Int, U, t; pbc=false)
-
-    Convenience constructor for a 1D nearest-neighbor Hubbard model with L sites.
-
-    """
-    function hubbard_hamiltonian(L::Int, U, t; pbc=false)
-        A = Matrix{Bool}(I, L, L)
-        pbc && (A[L,1] = A[1,L] = true)
-        return hubbard_hamiltonian(A, U, t)
-    end
-
-
-
-    """                
-        fullpauli(n::Int)
-            
-    The pool of all (4^n) n-qubit Pauli operators.
-
-    # Parameters
-    - `n`: Number of qubits in the system
-
-    # Returns
-    - `pool`: the full pauli pool.
-    """ 
-    function fullpauli(n::Int)
-        pool = ScaledPauliVector{n}[]
-        for plist in Iterators.product(ntuple(i->["I","X","Y","Z"],n)...)
-            pstr = join(plist)
-            pauli = [ScaledPauli(Pauli(pstr))]
-            push!(pool, pauli)
-        end
-        return pool
     end
 
     """                
@@ -303,6 +220,27 @@ module Pools
         return pool
     end
 
+    """                
+        fullpauli(n::Int)
+            
+    The pool of all (4^n) n-qubit Pauli operators.
+
+    # Parameters
+    - `n`: Number of qubits in the system
+
+    # Returns
+    - `pool`: the full pauli pool.
+    """ 
+    function fullpauli(n::Int)
+        pool = ScaledPauliVector{n}[]
+        for plist in Iterators.product(ntuple(i->["I","X","Y","Z"],n)...)
+            pstr = join(plist)
+            pauli = [ScaledPauli(Pauli(pstr))]
+            push!(pool, pauli)
+        end
+        return pool
+    end
+
     """
         one_local_pool(n::Int64, axes=["I","X","Y","Z"])
                                             
@@ -333,7 +271,7 @@ module Pools
     - `n`: Number of qubits in the system
 
     # Returns
-    - `pool`: the one-local pool.                                                
+    - `pool`: the two-local pool.                                                
     """                                                                                     
     function two_local_pool(n::Int64, axes=["X","Y","Z"])
         pool = ScaledPauliVector{n}[]
